A database over a socket says `working` is wrong by 55x, and the report now says so
Trial 4, and it did what a trial is for: it found a defect in a column shipped four
commits ago.

PostgreSQL 17 over a socket, 8 workers, fan-out 8, 20 s, 1,156 requests. Clock 201.4% of
nominal, no failed samples -- and it hardly matters, because the finding is a ratio
between two figures from the same run:

  CPU the process actually used              1.03 s   <- the OS, which sees native waits
  CPU the profiler attributes to requests   56.99 s   (working 2.85 x 17.30 ms x 1,156)
                                            55.26x

Java reports a thread inside a native call as RUNNABLE, and `working` is built on thread
state, so eight threads doing nothing but waiting for another process read as 2.85
threads on a CPU. The true figure is about 0.05. The fine tier says it more starkly:
`execute` holds 99.972% of the run at 6.25 ms per call and reads waiting 0.0%. The only
waiting detected anywhere was the caller parked in invokeAll -- inside 3.85 against
working 2.85, the difference of exactly 1.0.

`inside` is unaffected and correct: it counts threads in the execution whatever they were
doing. `working` and `waiting` are wrong together, in the same direction, by the same
amount.

The report already contained the contradiction. Its own duty-cycle header read "threads
were on CPU 0.63% of sampled wall time" twenty lines above a column that ignored it.

The fix is a bound, not a correction. `working` prints as 2.83/0.04 when the measured CPU
duty cycle cannot support it -- value over its ceiling, the idiom `in flight` already
uses -- with a warning naming each type. Both readings stated, because the column is
right on a CPU-bound operation and only the reader knows which they have.

The ceiling is inside x labelledDuty, from getThreadCpuTime, which sees through the JNI
boundary. It allows a factor of 1.5 because it applies a run-wide duty cycle to one
operation, and that factor is load-bearing rather than caution: Lucene reads working 6.40
against a ceiling of 6.40 and Calcite 1.00 against 0.97. A strict test would have accused
both of the defect PostgreSQL actually has.

Also: ideas.md item 24, measuring CPU per label instead of bounding thread state after
the fact. The bound says the number cannot be trusted; it does not produce a better one.

+++ trial-jdbc/src/main/kotlin/com/minogin/profiler/trial/jdbc/Pg.kt
@@ -165,9 +165,12 @@ fun ensureSchema(): Boolean {
                 """
                 INSERT INTO events (id, user_id, kind, amount, created_at, payload)
                 SELECT g,
-                       (g * 2654435761)::bigint % $USERS,
+                       -- g::bigint before the multiply, not after: generate_series yields int, and
+                       -- 5,000,000 * 2654435761 overflows int long before any cast on the result
+                       -- could rescue it. PostgreSQL says "integer out of range" and it is right.
+                       ((g::bigint * 2654435761) % $USERS)::int,
                        (g % 8)::smallint,
-                       ((g * 7919) % 100000) / 100.0,
+                       ((g::bigint * 7919) % 100000) / 100.0,
                        now() - ((g % 2592000) * interval '1 second'),
                        md5(g::text) || md5((g + 1)::text)
                 FROM generate_series(1, $ROWS) g
